requirements.txt
gkeepapi
Click==7.0
Flask==1.0.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10
jsonpickle==1.0
MarkupSafe==1.1.0
py-dateutil==2.2
six==1.11.0
Werkzeug==0.14.1
